+++ crates/json/src/lib.rs
@@ -1,2 +1,2 @@
-mod convert;
-pub mod parse;
+pub mod file;
+mod parse;
